Update src.py
Fixed a bug where GetForecastDates method would run for a specific number of times (240). Now it checks the length of the file that is used in the method and the number of iterations are determined by the length of this file.

Added some extra fancy factor when information is printed for the user.